Create helloWorld.py
Creating a code folder and doing some code practice. We can put code snippets here or delete the folder. I'm not sure good with understanding how people like GitHub documentation, but if this folder is allowed to be here then it'll be helpful to me; otherwise, we can delete it. :)